+++ src/main/java/com/example/pbl3_1/controller/dto/product/ProductForHomeDTO.java
@@ -10,14 +10,14 @@ import lombok.Setter;
 public class ProductForHomeDTO {
     private Long id;
     private String name;
-    private Float price;
-    private Float discount;
+    private Integer price;
+    private Integer discount;
     private String productImgPath;
     private Long sellerId;
     private String sellerName;
     private String sellerAvatar;
 
-    public ProductForHomeDTO(Long id, String name, Float price, Float discount, String productImgPath) {
+    public ProductForHomeDTO(Long id, String name, Integer price, Integer discount, String productImgPath) {
         this.id = id;
         this.name = name;
         this.price = price;
